port the user interface to English
For a wider audience of the program, the users' interface was
translated into English.  As checked with both reference files
test_in_01.txt and test_in_02.txt, file input, computation of a
solution, and reporting the results to the CLI and a permanent
record are not affected.  By `diff -s` vs `test_out_01.txt`, the
results of the solver are not modified either.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>