simplify some declarations
There were multiple declarations in pattern of `integer(1)` which
better are declared as `integer` only.  An accidental conversion
from INTEGER(4) to INTEGER(1), as the compilation with gfortran's
option -Wall describes as possible, then is less likely.  This
equally is a point suggested earlier by Vincent.

File I/O, validity test, provision of a solution for a grid read
from a file (check with diff -s, test_in_01.txt, test_out_02.txt)
do not reveal a change in the functionality.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>