halt the program if suggested input file is absent
Assume the program provided a randomly filled grid, followed by
the request to now read a grid from a file.  It is perceived as
to easy to misread in a hurry the previously displayed grid as
the new one already solved.  To prevent this error to occur, the
program now stops if requested to read a file wich is absent.

File I/O, validity test, compute a solution for a grid read from
a file (check with diff -s with test_in_02.txt and test_out_02.txt)
work as anticipated.  No change in the functionality.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>