application of fprettify ruled reformat
This commit combines the application of the fprettify rule
based reformat with additional manual edits (truncation of
long lines, etc.) to eventually ease reading of git's reflogs
and diff views.
Reference arrays in check.f90 were protected by "fprettify
fences" to retain some visual guide which does not affect the
result of fpm test, nor the diff test launched by the Makefile.
For the present project, with editors with syntax highlighting,
I'm not sure if an horizontal alignment of variable declarations
(and not only around `::`, as discussed on fprettify[1]) is of
significant advantage here.

[1] https://github.com/pseewald/fprettify/issues/157

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>